Setup selected day for the bar graph
Adjust font for the column date whenever it is the same date as the selected day.
Enable tapping on a bar date to move the selected day.
Fix a small problem with the scrolling so that when you add an element to the habit records, it centers the day that is selected