Extract a model that is actually Used to its own file
This model is needed because there was a crash whenever I was attempting to use access the same field in a SwiftData object - something with the relationship fault was killing the vibe. I tried a few different insertion methods of the data and nothing was solving this problem so instead, I transformed the SwiftData object into an analogous normal object and I was able to access the data fine.