Ensure that the chart is not empty whenever we are displaying the activity detail chart
There is a crash where if you have an activity that has never had a numeric detail logged and you go to the habit detail screen, edit the habit to have a detail. Then log the habit on the HabitDetailView, it will not update the `habitRecordsForDay`, it knows it should have a chart but nothing in the chart so it crashes. Simple fix is to ensure that it is not empty. But the better fix would be to update the records with each time we add a record so that the chart will appear accrdingly