Fixed little line that would randomly appear in the BlockStack
I am putting withAnimation in the View Model for easiness. This would ideally not be the case. I would rather have this be implemented via the view. Maybe performing the action instead. This is because in a real project I would want to allow the view model to be used as-is with any other view implementation (like UIKIt). And this muddies that a little bit. But it works.
Discovered that i was the transition with all the combines that made the little 1 pixel line in the VStack. This is a SwiftUI bug. I simplified the transition a little bit and it worked.