Removing explicit CalculationType - part 1
I am doing this because it doesn't add very much to the app but it makes it a lot more confusing. The only place that I believe this is being used is in the graph in habit details. And that could be done better to allow a selection of the type instead of being beholden to one type. It will be implicit for now